feat: Priority Sorting
For most of its existence, MIME::Types has supported a fairly complex
sorting algorithm that does not perform well at runtime especially with
a growing dataset of MIME types.

This branch is based on the idea that we can quantize the sorting of
similar types into a byte-packed bucket. For full details of the sort
mechanism, see Ruby MIME::Types 3.7.0.

The changes here provide the necessary updates to support the new field
(sort priority) and to do an in-place update of the data.