feat: Priority Sorting
This is the data half of priority sorting support.

for most of its existence, MIME::Types has supported a fairly complex
sorting algorithm that does not necessarily perform well at runtime
especially with a growing dataset of MIME types.

This experimental branch is based on the idea that we can quantize the
sorting of similar types into a bye-packed bucket. For full details of
the sort mechanism, see the corresponding experimental branch in
ruby-mime-types.

The changes here provide the necessary updates to support the new fields
(sort priority and extension priority) and to do an in-place update of
the data.